Fix build error: move demo components out of `pages` directory
- Next.js treats any `.{js|jsx|ts|tsx}` file under the `pages` directory as
   a page, and tries to build it as a page when we run `next build`
- This means you cannot store components within the `pages` directory
- Fix is simple: Move the components from `pages/lp` into `docs/components/DemoComponents`

I am able to start `yarn dev` and get the page to load, so I assume the (long) build will pass.
Also planning to take a look at the `ts-ignore` suppressions
